Revert "Change item `max_uses` and `remaining_uses` from u8 to u16"

(cherry picked from commit d28f14978b8f18266502278d97a4b91230499327)

+++ src/xrGame/ui/UICellItem.cpp
@@ -224,11 +224,11 @@ void CUICellItem::UpdateConditionProgressBar()
 			CEatableItem* eitm = smart_cast<CEatableItem*>(itm);
 			if (eitm)
 			{
-				u16 max_uses = eitm->GetMaxUses();
+				u8 max_uses = eitm->GetMaxUses();
 
 				if (max_uses > 0)
 				{
-					u16 remaining_uses = eitm->GetRemainingUses();
+					u8 remaining_uses = eitm->GetRemainingUses();
 
 					if (max_uses < 10)
 					{
